refactor: move functions into services
