+++ apps/dashboard/package.json
@@ -51,7 +51,6 @@
     "@openstatus/subscriptions": "workspace:*",
     "@openstatus/theme-store": "workspace:*",
     "@openstatus/tinybird": "workspace:*",
-    "@openstatus/tracker": "workspace:*",
     "@openstatus/ui": "workspace:*",
     "@openstatus/upstash": "workspace:*",
     "@openstatus/utils": "workspace:*",
